+++ validator-api/src/main/java/no/difi/vefa/validator/api/Document.java
@@ -1,20 +1,52 @@
 package no.difi.vefa.validator.api;
 
 import java.io.ByteArrayInputStream;
+import java.io.IOException;
 
-public interface Document {
+/**
+ * Representation of validation document.
+ */
+public class Document {
+
+    /**
+     * Document as #ByteArrayInputStream.
+     */
+    private ByteArrayInputStream byteArrayInputStream;
+    private String declaration;
+    private Expectation expectation;
+
+    @SuppressWarnings("all")
+    public Document(ByteArrayInputStream inputStream, String declaration, Expectation expectation) throws IOException, ValidatorException {
+        this.byteArrayInputStream = inputStream;
+        this.declaration = declaration;
+        this.expectation = expectation;
+    }
 
     /**
      * Declaration detected in document for validation.
      *
      * @return Declaration
      */
-    Declaration getDeclaration();
+    public String getDeclaration() {
+        return declaration;
+    }
+
+    /**
+     * Expectations detected in document for validation.
+     *
+     * @return Expectations
+     */
+    public Expectation getExpectation() {
+        return expectation;
+    }
 
     /**
      * Helper returning validated document as ByteArrayInputStream ready for use.
      *
      * @return Validated document
      */
-    ByteArrayInputStream getInputStream();
+    public ByteArrayInputStream getInputStream() {
+        byteArrayInputStream.reset();
+        return byteArrayInputStream;
+    }
 }