+++ validator-api/src/main/java/no/difi/vefa/validator/api/Expectation.java
@@ -1,0 +1,7 @@
+package no.difi.vefa.validator.api;
+
+public interface Expectation extends FlagFilterer {
+    String getDescription();
+
+    void verify(Section section);
+}